+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/VideoForwarder.kt
@@ -2,6 +2,7 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.forwarders
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeBlocking
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.textOrCaptionToMarkdown
 import com.pengrad.telegrambot.TelegramBot
 import com.pengrad.telegrambot.model.Message
 import com.pengrad.telegrambot.model.request.ParseMode
@@ -24,8 +25,12 @@ class VideoForwarder : Forwarder {
                 targetChatId,
                 message.video().fileId()
             ).apply {
-                message.caption() ?.let {
-                    caption(it)
+                message.textOrCaptionToMarkdown() ?.also {
+                    caption ->
+                    caption(caption)
+                    parseMode(
+                        ParseMode.Markdown
+                    )
                 }
                 parseMode(ParseMode.Markdown)
             }