+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/VoiceForwarder.kt
@@ -2,8 +2,10 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.forwarders
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeBlocking
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.textOrCaptionToMarkdown
 import com.pengrad.telegrambot.TelegramBot
 import com.pengrad.telegrambot.model.Message
+import com.pengrad.telegrambot.model.request.ParseMode
 import com.pengrad.telegrambot.request.SendVoice
 import java.io.IOException
 
@@ -23,8 +25,12 @@ class VoiceForwarder : Forwarder {
                 targetChatId,
                 message.voice().fileId()
             ).apply {
-                message.caption() ?.let {
-                    caption(it)
+                message.textOrCaptionToMarkdown() ?.also {
+                    caption ->
+                    caption(caption)
+                    parseMode(
+                        ParseMode.Markdown
+                    )
                 }
                 message.voice().also {
                     it.duration() ?.let {